refactor: move methods into digital credential service
