+++ legal-api/src/legal_api/services/digital_credentials.py
@@ -17,17 +17,32 @@
 
 import json
 from contextlib import suppress
+from enum import Enum
 from typing import Optional
 
 import requests
 
 from legal_api.decorators import requires_traction_auth
-from legal_api.helpers.digital_credentials import DCRevocationReason
-from legal_api.models import DCDefinition
+from legal_api.models import Business, CorpType, DCDefinition, DCIssuedBusinessUserCredential, User
 
 
 class DigitalCredentialsService:
     """Provides services to do digital credentials using aca-py agent."""
+
+    class DCRevocationReason(Enum):
+        """Digital Credential Revocation Reasons."""
+
+        UPDATED_INFORMATION = 'You were offered a new credential with updated information \
+            and that revoked all previous copies.'
+        VOLUNTARY_DISSOLUTION = 'You chose to dissolve your business. \
+            A new credential was offered that reflects the new company status and that revoked all previous copies.'
+        ADMINISTRATIVE_DISSOLUTION = 'Your business was dissolved by the Registrar.'
+        PUT_BACK_ON = 'Your business was put back on the Registry.'
+        RESTORATION = 'Your business was put back on the Registry. \
+            A new credential was offered that reflects the new company status and that revoked all previous copies.'
+        ACCESS_REMOVED = 'Your role in the business was changed and you no longer have system access to the business.'
+        SELF_REISSUANCE = 'You chose to issue yourself a new credential and that revoked all previous copies.'
+        SELF_REVOCATION = 'You chose to revoke your own credential.'
 
     def __init__(self):
         """Initialize this object."""
@@ -259,3 +274,84 @@ class DigitalCredentialsService:
             'Content-Type': 'application/json',
             'Authorization': f'Bearer {self.app.api_token}'
         }
+
+    def get_digital_credential_data(self, business: Business, user: User, credential_type: DCDefinition.CredentialType):
+        """Get the data for a digital credential."""
+        if credential_type == DCDefinition.CredentialType.business:
+
+            # Find the credential id from dc_issued_business_user_credentials and if there isn't one create one
+            if not (issued_business_user_credential := DCIssuedBusinessUserCredential.find_by(
+                    business_id=business.id, user_id=user.id)):
+                issued_business_user_credential = DCIssuedBusinessUserCredential(
+                    business_id=business.id, user_id=user.id)
+                issued_business_user_credential.save()
+
+            credential_id = f'{issued_business_user_credential.id:08}'
+
+            if (business_type := CorpType.find_by_id(business.legal_type)):
+                business_type = business_type.full_desc
+            else:
+                business_type = business.legal_type
+
+            registered_on_dateint = ''
+            if business.founding_date:
+                registered_on_dateint = business.founding_date.strftime('%Y%m%d')
+
+            company_status = Business.State(business.state).name
+
+            family_name = (user.lastname or '').strip().upper()
+
+            given_names = ' '.join([x.strip() for x in [user.firstname, user.middlename] if x and x.strip()]).upper()
+
+            return [
+                {
+                    'name': 'credential_id',
+                    'value':  credential_id or ''
+                },
+                {
+                    'name': 'identifier',
+                    'value': business.identifier or ''
+                },
+                {
+                    'name': 'business_name',
+                    'value': business.legal_name or ''
+                },
+                {
+                    'name': 'business_type',
+                    'value': business_type or ''
+                },
+                {
+                    'name': 'cra_business_number',
+                    'value': business.tax_id or ''
+                },
+                {
+                    'name': 'registered_on_dateint',
+                    'value': registered_on_dateint or ''
+                },
+                {
+                    'name': 'company_status',
+                    'value': company_status or ''
+                },
+                {
+                    'name': 'family_name',
+                    'value': family_name or ''
+                },
+                {
+                    'name': 'given_names',
+                    'value': given_names or ''
+                },
+                {
+                    'name': 'role',
+                    'value': ''
+                }
+            ]
+
+        return None
+
+    def extract_invitation_message_id(self, json_message: dict):
+        """Extract the invitation message id from the json message."""
+        if 'invitation' in json_message and json_message['invitation'] is not None:
+            invitation_message_id = json_message['invitation']['@id']
+        else:
+            invitation_message_id = json_message['invitation_msg_id']
+        return invitation_message_id
