+++ queue_services/entity-digital-credentials/requirements.txt
@@ -1,7 +1,7 @@
 aiohttp
 asyncio-nats-streaming
 asyncio-nats-client==0.11.4
-attrs==20.3.0
+attrs==23.1.0
 blinker==1.4
 certifi==2020.12.5
 click==8.1.3
@@ -9,7 +9,7 @@ dpath==2.0.1
 Flask==1.1.2
 Jinja2==2.11.3
 itsdangerous==1.1.0
-jsonschema==4.16.0
+jsonschema==4.19.0
 MarkupSafe==1.1.1
 protobuf==3.15.8
 pyrsistent==0.17.3
